Split string helpers from pyanaconda.core.util
