+++ pyanaconda/modules/payloads/payload/dnf/utils.py
@@ -26,7 +26,8 @@ from pyanaconda.anaconda_loggers import get_module_logger
 from pyanaconda.core import util
 from pyanaconda.core.configuration.anaconda import conf
 from pyanaconda.core.regexes import VERSION_DIGITS
-from pyanaconda.core.util import is_lpae_available, decode_bytes, join_paths, execWithCapture
+from pyanaconda.core.util import is_lpae_available, join_paths, execWithCapture
+from pyanaconda.core.string import decode_bytes
 from pyanaconda.modules.common.constants.objects import DEVICE_TREE
 from pyanaconda.modules.common.constants.services import STORAGE
 from pyanaconda.modules.common.structures.packages import PackagesSelectionData
