+++ pyanaconda/modules/storage/bootloader/utils.py
@@ -21,7 +21,8 @@ from glob import glob
 from pyanaconda.modules.common.errors.installation import BootloaderInstallationError
 from pyanaconda.modules.storage.bootloader.image import LinuxBootLoaderImage
 from pyanaconda.core.configuration.anaconda import conf
-from pyanaconda.core.util import decode_bytes, execWithRedirect
+from pyanaconda.core.util import execWithRedirect
+from pyanaconda.core.string import decode_bytes
 from pyanaconda.product import productName
 
 from pyanaconda.anaconda_loggers import get_module_logger
