+++ pyanaconda/queuefactory.py
@@ -17,7 +17,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import queue
-from pyanaconda.core.util import lowerASCII, upperASCII
+from pyanaconda.core.string import upperASCII, lowerASCII
 
 
 class QueueFactory(object):
