+++ pyanaconda/ui/gui/spokes/lib/custom_storage_helpers.py
@@ -25,7 +25,7 @@ from pyanaconda.core.constants import SIZE_UNITS_DEFAULT
 from pyanaconda.core.i18n import _, N_, CN_, C_
 from pyanaconda.core.storage import PROTECTED_FORMAT_TYPES, SIZE_POLICY_AUTO, SIZE_POLICY_MAX, \
     DEVICE_TYPE_LVM, DEVICE_TYPE_BTRFS, DEVICE_TYPE_LVM_THINP, DEVICE_TYPE_MD, Size
-from pyanaconda.core.util import lowerASCII
+from pyanaconda.core.string import lowerASCII
 from pyanaconda.modules.common.structures.device_factory import DeviceFactoryRequest, \
     DeviceFactoryPermissions
 from pyanaconda.modules.common.structures.storage import DeviceFormatData, DeviceData
